+++ lib/stone_bank_web/controllers/user_controller.ex
@@ -1,6 +1,7 @@
 defmodule StoneBankWeb.UserController do
   use StoneBankWeb, :controller
   alias StoneBank.Accounts
+  alias StoneBank.Accounts.Auth.Guardian
 
   action_fallback StoneBankWeb.FallbackController
 
@@ -10,6 +11,14 @@ defmodule StoneBankWeb.UserController do
       |> put_status(:created)
       |> put_resp_header("location", Routes.user_path(conn, :show, id: user.id))
       |> render("account.json", %{user: user, account: account})
+    end
+  end
+
+  def signin(conn, %{"email" => email, "password" => password}) do
+    with {:ok, user, token} <- Guardian.authenticate(email, password) do
+      conn
+      |> put_status(:created)
+      |> render("user_auth.json", user: user, token: token)
     end
   end