+++ lib/stone_bank_web/views/user_view.ex
@@ -23,6 +23,11 @@ defmodule StoneBankWeb.UserView do
     %{data: render_many(users, __MODULE__, "user.json")}
   end
 
+  def render("user_auth.json", %{user: user, token: token}) do
+    user = Map.put(render_one(user, __MODULE__, "user.json"), :token, token)
+    %{data: user}
+  end
+
   def render("user.json", %{user: user}) do
     %{
       id: user.id,