+++ CHANGELOG.md
@@ -5,9 +5,9 @@
 ## [0.0.22] - 2021-04-10
 ### Added
 
-- add guardian dependency
-- add guardian and session module for jwt authenticate
-- add guardian config
+- add signin route
+- add signin def at user_controller
+- add user_auth for json response
 
 ### Changed